Fix one small warning on slc4, LumiIntegrateFSR.cpp line 542,  comparison between signed and unsigned ints
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@115389 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 3b4a81f4acabdb4a93b3d60bb271548412f2dd27

+++ Rec/Rec/LumiAlgs/src/LumiIntegrateFSR.cpp
@@ -538,8 +538,8 @@ StatusCode LumiIntegrateFSR::trigger_event( std::string primaryFileRecordAddress
     // pick the TimeSpanFSR
     LHCb::TimeSpanFSRs::iterator tsfsr = timeSpanFSRs->begin();
     // check index bounds
-    long tsfsr_len = timeSpanFSRs->size();
-    if ( fkey > tsfsr_len-1 ) {
+    unsigned long tsfsr_len = timeSpanFSRs->size();
+    if ( (unsigned long)fkey > tsfsr_len-1 ) { 
       warning() << "missing timeSpanFSR - use previous DB conditions" << endmsg;
       return StatusCode::SUCCESS;
     }
